requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2026.1.4
python-dotenv==1.0.1
python-telegram-bot==13.15
pytz==2025.2
ruff==0.15.1
six==1.17.0
tornado==6.1
tzdata==2025.3
tzlocal==5.3.1
